shut up warnings on NetBSD and FreeBSD, pointed out by Soda-san (n-soda@)
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@5840 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ bench/gfiops/gfiops.c
@@ -11,6 +11,7 @@
 #include <sys/wait.h>
 #include <pthread.h>
 #include <errno.h>
+#include <libgen.h>
 
 #include <gfarm/gfarm.h>
 
@@ -521,8 +522,8 @@ result(const char *name, struct timeval *t1, struct timeval *t2,
 
 	printf("%9s: %6d /sec: %d proc * %4d: "
 	       "%2ld.%06ld sec: %9.3f usec * %d\n",
-	       name, num_persec, nprocs, ntimes, t2->tv_sec, t2->tv_usec,
-	       usec, n);
+	       name, num_persec, nprocs, ntimes,
+	       (long int) t2->tv_sec, t2->tv_usec, usec, n);
 }
 
 static int
